mark as available works
