+++ src/lib/components/Calendar/ScheduleTable.svelte
@@ -1,25 +1,20 @@
 <script lang="ts">
-	import Button from '$lib/components/Button.svelte';
-	import Legend from '$lib/components/Legend.svelte';
 	import { EMOTICONS_REVERSE } from '$lib/constants';
 	import {
-		closeEditor,
 		getRowColor,
 		isAvailableOnRow,
-		markRowAsAvailable,
 		markRowUnavailableLocally,
-		requestToMarkOneRow,
-		showEditor,
-		toggleEmoticon
+		requestToMarkOneRow
 	} from '$lib/logics/Calendar/ScheduleTable/logic';
-	import { EMOTICONS, UNAVAILABLE } from '$lib/logics/Calendar/_shared/constants';
 	import type { Row, Unavailable } from '$lib/types';
 	import { AvailabilityStatus } from '@prisma/client';
-	import { createEventDispatcher } from 'svelte';
+	import { createEventDispatcher, tick } from 'svelte';
+	import Editor from './Editor.svelte';
 
 	const dispatch = createEventDispatcher();
 
-	export let rows: Row[];
+	export let dbRows: Row[];
+	export let displayedRows: Row[];
 	export let timeZone: string;
 
 	let openedRows = new Set<number>(); // row inds which have open editors
@@ -34,69 +29,72 @@
 	let rowColors: string[] = [];
 
 	$: {
-		if (!rows.length) break $;
+		if (!dbRows.length) break $;
 
 		rowColors = [...Array(21).keys()].map((_, i) =>
 			getRowColor({
 				i,
-				numRows: rows.length,
-				isAvailable: isAvailableOnRow(rows[i]),
+				numRows: dbRows.length,
+				isAvailable: isAvailableOnRow(dbRows[i]),
 				isRowExpanded: openedRows.has(i)
 			})
 		);
 	}
 
-	const handleEditorOpen = (i: number) => {
-		openedRows = showEditor({ i, openedRows });
+	const openEditor = (i: number) => {
+		openedRows.add(i);
+		openedRows = openedRows;
+		tick().then(() => {
+			document.getElementById(`editor-${i}`)?.focus();
+		});
+	};
+
+	const closeEditor = (i: number) => {
+		openedRows.delete(i);
+		openedRows = openedRows;
 	};
 
 	const markRowAsUnavailable = async ({ i, status }: { i: number; status: Unavailable }) => {
-		const dbRow = { ...rows[i] };
-		rows = markRowUnavailableLocally({ i, displayedRows: rows, status });
-		dispatch('changed:rows', rows);
+		const dbRow = { ...displayedRows[i] };
+		displayedRows = markRowUnavailableLocally({ i, displayedRows, status });
+		dispatch('changed:displayedRows', displayedRows);
 		try {
 			await requestToMarkOneRow({
-				i,
 				status,
-				displayedRows: rows,
+				displayedRow: displayedRows[i],
 				availableDetails: null
 			});
-			closeEditor({ i, openedRows });
+			closeEditor(i);
 			dispatch('markedRow');
 		} catch (err) {
 			alert('Something went wrong with saving'); // TODO: come up with better UI for showing err
 			console.error('Something went wrong with marking row as unavailable', err);
-			rows[i] = dbRow;
-			dispatch('changed:rows', rows);
+			displayedRows[i] = dbRow;
+			dispatch('changed:displayedRows', displayedRows);
 		}
 	};
 </script>
 
 <table id="schedule">
-	{#each rows as row, i}
+	{#each dbRows as row, i}
 		<tr style="background-color: {rowColors[i]};">
 			<td
 				class:blue={openedRows.has(i)}
 				class="day"
-				on:click={() => handleEditorOpen(i)}
-				on:keyup={() => handleEditorOpen(i)}
+				on:click={() => openEditor(i)}
+				on:keyup={() => openEditor(i)}
 			>
 				{row.englishDay}
 			</td>
 			<td
 				class:blue={openedRows.has(i)}
 				class="date"
-				on:click={() => handleEditorOpen(i)}
-				on:keyup={() => handleEditorOpen(i)}
+				on:click={() => openEditor(i)}
+				on:keyup={() => openEditor(i)}
 			>
 				{row.monthDay}
 			</td>
-			<td
-				colspan="2"
-				class="time"
-				on:click={() => handleEditorOpen(i)}
-				on:keyup={() => handleEditorOpen(i)}
-			>
+			<td colspan="2" class="time" on:click={() => openEditor(i)} on:keyup={() => openEditor(i)}>
 				{#if row.availRange === AvailabilityStatus.UNSPECIFIED}
 					<p>Unspecified (<span class="edit">edit</span>)</p>
 				{:else if row.availRange === AvailabilityStatus.BUSY}
@@ -152,105 +150,33 @@
 			{/if}
 		</tr>
 		{#if openedRows.has(i)}
-			<tr style="background: #A0E3FF">
-				<td colspan="5" style="padding: 0.9rem 0.4rem;" class="editorCell">
-					<form on:submit|preventDefault={() => {}}>
-						<div class="v-center-h-space flex-col" style="gap: 0.1rem;">
-							<!-- prettier-ignore -->
-							<input
-                id={`editor-${i}`}
-                type="text"
-                class="text-inherit"
-                placeholder='Enter a valid time range. Ex. "2:30pm-7 or 5-6"'
-                value={UNAVAILABLE.includes(rows[i].availRange) ? '' : rows[i].availRange}
-                on:keydown={(e) => {
-                  rowIndsWithTimeErrs.delete(i);
-                  rowIndsWithTimeErrs = new Set(rowIndsWithTimeErrs);
-
-									rows[i].availRange = e.currentTarget.value;
-                }}
-              />
-							{#if rowIndsWithTimeErrs.has(i)}
-								<p class="red">Enter a valid time range. Ex. "2:30pm-7 or 5-6"</p>
-							{/if}
-						</div>
-						<div class="v-center-h-space">
-							{#key rows[i].emoticons}
-								{#each Object.entries(EMOTICONS) as [emoji, emojiDescrip]}
-									<div
-										class="emoji {rows[i].emoticons.has(emojiDescrip) ? 'chosen' : ''}"
-										on:click={() =>
-											toggleEmoticon({ i, displayedRows: rows, emoticon: emojiDescrip })}
-										on:keyup={() =>
-											toggleEmoticon({ i, displayedRows: rows, emoticon: emojiDescrip })}
-									>
-										{emoji}
-									</div>
-								{/each}
-							{/key}
-							<div class="tooltip">
-								<p>
-									Legend
-									<span class="tooltiptext">
-										<Legend />
-									</span>
-								</p>
-							</div>
-						</div>
-						<div class="v-center-h-space">
-							<textarea
-								bind:value={rows[i].notes}
-								class="text-inherit"
-								name="notes"
-								placeholder="(add notes)"
-							/>
-						</div>
-						<div class="editor-btns">
-							<Button
-								onClick={async () => {
-									try {
-										await markRowAsAvailable({
-											i,
-											displayedRows: rows,
-											timeZone,
-											rowIndsWithTimeErrs
-										});
-										openedRows = closeEditor({ i, openedRows });
-									} catch {}
-								}}
-								disabled={rowIndsWithTimeErrs.has(i)}
-								content={'Save'}
-								bgColor={'#93FF8B'}
-								padding="0.1rem 0.7rem"
-								color={'black'}
-								fontSize={'larger'}
-							/>
-							<Button
-								onClick={(e) => {
-									e?.preventDefault();
-									openedRows = closeEditor({ i, openedRows });
-								}}
-								content={'Cancel'}
-								bgColor={'rgba(255, 233, 184, 0.78)'}
-								padding="0.1rem 0.7rem"
-								color={'black'}
-								fontSize={'larger'}
-							/>
-						</div>
-					</form>
-				</td>
-			</tr>
+			<Editor
+				hasBadTime={rowIndsWithTimeErrs.has(i)}
+				unsavedRows={displayedRows}
+				{i}
+				{timeZone}
+				on:changed:time={(e) => {
+					rowIndsWithTimeErrs.delete(i);
+					rowIndsWithTimeErrs = rowIndsWithTimeErrs;
+					dispatch('changed:displayedRow', {
+						i,
+						row: { ...displayedRows[i], availRange: e.detail }
+					});
+				}}
+				on:closeEditor={() => closeEditor(i)}
+				on:clicked:emoji={(e) => {
+					dispatch('changed:displayedRow', {
+						i,
+						row: { ...displayedRows[i], emoticons: e.detail }
+					});
+				}}
+				on:markedRow={() => dispatch('markedRow')}
+			/>
 		{/if}
 	{/each}
 </table>
 
 <style>
-	.red {
-		color: #bd0000;
-	}
-	.text-inherit {
-		font-size: inherit;
-	}
 	table {
 		border-collapse: collapse;
 		border: 1px solid #dddddd;
@@ -303,49 +229,7 @@
 		cursor: pointer;
 	}
 
-	#schedule td.editorCell {
-		width: 100%;
-	}
-	.editor-btns {
-		gap: 0.5rem;
-		display: flex;
-		justify-content: center;
-		margin: 1rem auto 0;
-	}
 	.blue {
 		background: #a0e3ff;
-	}
-
-	.emoji.chosen {
-		border: 1px solid black;
-	}
-	textarea {
-		margin: 0;
-		resize: none;
-	}
-	.v-center-h-space {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-	}
-	.flex-col {
-		flex-direction: column;
-	}
-	.tooltip {
-		width: fit-content;
-	}
-
-	.emoji {
-		background: white;
-		font-size: x-large;
-		margin: 0.5rem 0;
-		padding-left: 0.2rem;
-		padding-right: 0.2rem;
-		border-radius: 3px;
-		border: 1px solid #d9d9d9;
-	}
-
-	.tooltiptext {
-		left: 0;
 	}
 </style>
